Made x11 connection example.

+++ src/libraries/linux/implementations/send_buffer.cpp
@@ -4,7 +4,7 @@
 
 auto nix::send_buffer(FileDescriptor const socket_descriptor,
                       void const* p_message_buffer, ssize const buffer_length,
-                      int4 const flags, Socket const* p_destination_socket,
+                      int8 const flags, Socket const* p_destination_socket,
                       ssize const addr_length) -> Result<ssize> {
     return nix::syscall6(44u, socket_descriptor, p_message_buffer,
                          buffer_length, flags, p_destination_socket,
